graph: Finish renaming `frecency_cache` to `lfu_cache`
