+++ graph/src/components/subgraph/instance.rs
@@ -1,5 +1,5 @@
 use crate::prelude::*;
-use crate::util::frecency_cache::LfuCache;
+use crate::util::lfu_cache::LfuCache;
 use web3::types::Log;
 
 #[derive(Clone, Debug)]
@@ -16,9 +16,9 @@ pub struct BlockState {
 }
 
 impl BlockState {
-    pub fn with_cache(frecency_cache: LfuCache<EntityKey, Option<Entity>>) -> Self {
+    pub fn with_cache(lfu_cache: LfuCache<EntityKey, Option<Entity>>) -> Self {
         BlockState {
-            entity_cache: EntityCache::with_current(frecency_cache),
+            entity_cache: EntityCache::with_current(lfu_cache),
             created_data_sources: Vec::new(),
         }
     }
